+++ custom/models/sale_order.py
@@ -1,6 +1,7 @@
 import logging
 from odoo import models, fields, api
 from odoo.tools import float_round
+from datetime import timedelta
 _logger = logging.getLogger(__name__)
 
 
@@ -30,12 +31,11 @@ class SaleOrder(models.Model):
         
     so_type_camion = fields.Selection(
         [
-            ('Fourgon 20m3 (150€ + 0.50€/km)','Fourgon 20m3 (150€ + 0.50€/km)'),
-            ('GEODIS','GEODIS'),
-            ('Porteur avec hayon (base)','Porteur avec hayon (base)'),
             ('Semi-remorque (base)','Semi-remorque (base)'),
             ('Semi-remorque avec hayon (base)','Semi-remorque avec hayon (base)'),
             ('Semi-remorque plateau (base)','Semi-remorque plateau (base)'),
+            ('Porteur avec hayon (base)','Porteur avec hayon (base)'),
+            ('Fourgon 20m3 (150€ + 0.50€/km)','Fourgon 20m3 (150€ + 0.50€/km)'),
             ('Semi-remorque chariot embarqué (650€)','Semi-remorque chariot embarqué (650€)'),
             ('Autre (sur devis)','Autre (sur devis)'),
         ],
@@ -58,7 +58,7 @@ class SaleOrder(models.Model):
     #@api.constrains('partner_id', 'so_commande_client')
     #def _check_commande_client_required(self):
         #for order in self:
-            # Liste des mots-clés qui nécessitent un N° Commande Client obligatoire P
+            # Liste des mots-clés qui nécessitent un N° Commande Client obligatoire
             #required_keywords = ['GCC', 'BOLLORE', 'BOUYGUES', 'LEGENDRE']
             
             # Si le nom du client contient un des mots-clés
@@ -75,7 +75,8 @@ class SaleOrder(models.Model):
     so_date_de_modification = fields.Date(string="Date de modification")
     so_date_de_commande = fields.Date(string="Date de la commande")
     so_date_bpe = fields.Date(string="BPE du : ")
-    so_date_de_reception_devis = fields.Date(string="Demande reçue le :", required=True)
+    so_date_de_reception_devis = fields.Date(
+    string="Demande reçue le :")
     so_date_du_devis = fields.Date(string="Devis fait le : ")
     so_date_de_modification_devis = fields.Date(string="Devis modifié le : ")
     so_date_devis_valide = fields.Date(string="Devis validé le : ")
@@ -102,7 +103,6 @@ class SaleOrder(models.Model):
             ('WICONA','WICONA'),
         ],
         string="GAMME",
-        required=True,
     )
 
     so_serie = fields.Selection(
@@ -230,7 +230,6 @@ class SaleOrder(models.Model):
             ('WT50.RC3', 'WT50.RC3'),
         ],
         string="SERIE",
-        required=True,
     )
 
 
@@ -349,4 +348,99 @@ class SaleOrder(models.Model):
         invoice_vals['x_studio_date_de_la_commande'] = self.x_studio_date_de_la_commande
         return invoice_vals
 
-    
+    #Below methods are shifted from fma_sale_order_custom module
+
+    # Init date de livraison prévue et synchronisation avec commitment_date
+    @api.depends('so_date_bpe', 'so_delai_confirme_en_semaine')
+    def _compute_so_date_de_livraison(self):
+        for order in self:
+            if order.so_date_bpe and order.so_delai_confirme_en_semaine:
+                # Calculer la date de livraison prévue
+                order.so_date_de_livraison = order.so_date_bpe + timedelta(weeks=order.so_delai_confirme_en_semaine)
+                # Synchroniser avec commitment_date
+                order.commitment_date = order.so_date_de_livraison
+            else:
+                # Réinitialiser si les valeurs nécessaires sont manquantes
+                order.so_date_de_livraison = False
+                order.commitment_date = False
+
+    # Calcul des marges et coûts pour le devis
+    @api.depends('so_mtt_facturer_devis', 'so_achat_vitrage_devis', 'so_achat_matiere_devis')
+    def _compute_so_marge_brute_devis(self):
+        for order in self:
+            order.so_marge_brute_devis = order.so_mtt_facturer_devis - order.so_achat_vitrage_devis - order.so_achat_matiere_devis
+
+    @api.depends('so_marge_brute_devis', 'so_cout_mod_devis')
+    def _compute_so_mcv_devis(self):
+        for order in self:
+            order.so_mcv_devis = order.so_marge_brute_devis - order.so_cout_mod_devis
+
+    # Calcul des marges et coûts pour BE
+    @api.depends('so_mtt_facturer_be', 'so_achat_vitrage_be', 'so_achat_matiere_be')
+    def _compute_so_marge_brute_be(self):
+        for order in self:
+            order.so_marge_brute_be = order.so_mtt_facturer_be - order.so_achat_vitrage_be - order.so_achat_matiere_be
+
+    @api.depends('so_marge_brute_be', 'so_cout_mod_be')
+    def _compute_so_mcv_be(self):
+        for order in self:
+            order.so_mcv_be = order.so_marge_brute_be - order.so_cout_mod_be
+
+    # Calcul des marges et coûts pour le réel
+    @api.depends('so_mtt_facturer_reel', 'so_achat_vitrage_reel', 'so_achat_matiere_reel')
+    def _compute_so_marge_brute_reel(self):
+        for order in self:
+            order.so_marge_brute_reel = order.so_mtt_facturer_reel - order.so_achat_vitrage_reel - order.so_achat_matiere_reel
+
+    @api.depends('so_marge_brute_reel', 'so_cout_mod_reel')
+    def _compute_so_mcv_reel(self):
+        for order in self:
+            order.so_mcv_reel = order.so_marge_brute_reel - order.so_cout_mod_reel
+
+    @api.depends('so_marge_brute_devis', 'so_mtt_facturer_devis')
+    def _compute_so_prc_marge_brute_devis(self):
+        for order in self:
+            if order.so_mtt_facturer_devis:
+                order.so_prc_marge_brute_devis = (order.so_marge_brute_devis / order.so_mtt_facturer_devis) * 100
+            else:
+                order.so_prc_marge_brute_devis = 0.0
+
+    @api.depends('so_mcv_devis', 'so_mtt_facturer_devis')
+    def _compute_so_prc_mcv_devis(self):
+        for order in self:
+            if order.so_mtt_facturer_devis:
+                order.so_prc_mcv_devis = (order.so_mcv_devis / order.so_mtt_facturer_devis) * 100
+            else:
+                order.so_prc_mcv_devis = 0.0
+
+    @api.depends('so_marge_brute_be', 'so_mtt_facturer_be')
+    def _compute_so_prc_marge_brute_be(self):
+        for order in self:
+            if order.so_mtt_facturer_be:
+                order.so_prc_marge_brute_be = (order.so_marge_brute_be / order.so_mtt_facturer_be) * 100
+            else:
+                order.so_prc_marge_brute_be = 0.0
+
+    @api.depends('so_mcv_be', 'so_mtt_facturer_be')
+    def _compute_so_prc_mcv_be(self):
+        for order in self:
+            if order.so_mtt_facturer_be:
+                order.so_prc_mcv_be = (order.so_mcv_be / order.so_mtt_facturer_be) * 100
+            else:
+                order.so_prc_mcv_be = 0.0
+
+    @api.depends('so_marge_brute_reel', 'so_mtt_facturer_reel')
+    def _compute_so_prc_marge_brute_reel(self):
+        for order in self:
+            if order.so_mtt_facturer_reel:
+                order.so_prc_marge_brute_reel = (order.so_marge_brute_reel / order.so_mtt_facturer_reel) * 100
+            else:
+                order.so_prc_marge_brute_reel = 0.0
+
+    @api.depends('so_mcv_reel', 'so_mtt_facturer_reel')
+    def _compute_so_prc_mcv_reel(self):
+        for order in self:
+            if order.so_mtt_facturer_reel:
+                order.so_prc_mcv_reel = (order.so_mcv_reel / order.so_mtt_facturer_reel) * 100
+            else:
+                order.so_prc_mcv_reel = 0.0
