+++ custom_colisage/models/colisage.py
@@ -1,5 +1,8 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 import logging
-from odoo import models, fields, api, re
+from odoo import models, fields, api
 
 class SaleOrder(models.Model):
     _inherit = ['sale.order', 'delivery']
@@ -11,6 +14,6 @@ class SaleOrder(models.Model):
     
 
 
-    line.product_id.name = re.sub(r'\[.*?\]', '', line.name).strip()
+    # line.product_id.name = re.sub(r'\[.*?\]', '', line.name).strip()
