+++ custom_delivery/__manifest__.py
@@ -14,10 +14,10 @@
     'website': "http://www.yourcompany.com",
 
     'category': 'Uncategorized',
-    'version': '0.1',
+    'version': '17.0.1.0.1',
 
     # any module necessary for this one to work correctly
-    'depends': ['base', 'stock'],
+    'depends': ['base', 'stock', 'custom'],
 
     # always loaded
     'data': [
