+++ custom_delivery/models/delivery.py
@@ -1,4 +1,4 @@
-from odoo import models, fields, api, re
+from odoo import models, fields, api
 
 class SaleOrder(models.Model):
     _inherit = ['delivery', 'sale.order']
