+++ custom_delivery/models/stock_picking.py
@@ -5,9 +5,10 @@ _logger = logging.getLogger(__name__)
 
 class StockPicking(models.Model):
     _inherit = 'stock.picking'
+    
 
-    def __init__(self, pool, cr):
-        super(StockPicking, self).__init__(pool, cr)
+    def __init__(self, env, ids, prefetch_ids):
+        super(StockPicking, self).__init__(env, ids, prefetch_ids)
         _logger.warning("Le modèle StockPicking est chargé avec succès")
 
     # Champ sale_id (s'il n'existe pas déjà)
@@ -17,12 +18,12 @@ class StockPicking(models.Model):
     so_horaire_ouverture_bl = fields.Float(string='Horaire ouverture', widget='float_time')
     so_horaire_fermeture_bl = fields.Float(string='Horaire fermeture', widget='float_time')
 
+
     # Forcer le recalcul après la modification des mouvements
     def write(self, vals):
         res = super(StockPicking, self).write(vals)
         if 'move_ids_without_package' in vals:  # Si les mouvements sont modifiés
-            # self._compute_reliquat_qty()  # Recalcul du reliquat (commenté car non défini)
-            _logger.warning("Appel à '_compute_reliquat_qty' commenté car la méthode n'est pas définie.")
+            self._compute_reliquat_qty()  # Recalcul du reliquat
         if 'scheduled_date' in vals:  # Mise à jour de la date de livraison si modifiée
             for picking in self:
                 if picking.sale_id:
@@ -31,14 +32,13 @@ class StockPicking(models.Model):
 
     so_type_camion_bl = fields.Selection(
         [
-            ('Fourgon 20m3 (150€ + 0.50€/km)','Fourgon 20m3 (150€ + 0.50€/km)'),
-            ('GEODIS','GEODIS'),
-            ('Porteur avec hayon (base)','Porteur avec hayon (base)'),
-            ('Semi-remorque (base)','Semi-remorque (base)'),
-            ('Semi-remorque avec hayon (base)','Semi-remorque avec hayon (base)'),
-            ('Semi-remorque plateau (base)','Semi-remorque plateau (base)'),
-            ('Semi-remorque chariot embarqué (650€)','Semi-remorque chariot embarqué (650€)'),
-            ('Autre (sur devis)','Autre (sur devis)'),
+            ('Semi-remorque (base)', 'Semi-remorque (base)'),
+            ('Semi-remorque avec hayon (base)', 'Semi-remorque avec hayon (base)'),
+            ('Semi-remorque plateau (base)', 'Semi-remorque plateau (base)'),
+            ('Porteur avec hayon (base)', 'Porteur avec hayon (base)'),
+            ('Fourgon 20m3 (150€ + 0.50€/km)', 'Fourgon 20m3 (150€ + 0.50€/km)'),
+            ('Semi-remorque chariot embarqué (650€)', 'Semi-remorque chariot embarqué (650€)'),
+            ('Autre (sur devis)', 'Autre (sur devis)'),
         ],
         string="Type de camion (Hayon palette maxi 2400mm)",
     )
@@ -73,9 +73,9 @@ class PickingColisageLine(models.Model):
     so_qte_livree = fields.Integer(string="Qté Livrée", track_visibility='onchange')
 
     # Contrainte SQL pour garantir l'unicité du champ 'so_repere'
-    #_sql_constraints = [
-    #    ('so_repere_unique', 'UNIQUE(so_repere)', 'La référence doit être unique !'),
-    #]
+    _sql_constraints = [
+        ('so_repere_unique', 'UNIQUE(so_repere)', 'La référence doit être unique !'),
+    ]
 
     @api.model
     def create(self, vals):
