+++ custom_invoice/__manifest__.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 {
     'name': 'Custom Invoice Text Block',
-    'version': '1.0',
+    'version': '17.0.1.0.1',
     'summary': 'Show text block on invoice based on contact boolean field',
     'author': 'Your Name',
-    'depends': ['account'],
+    'depends': ['account', 'custom'],
     'data': [
          'views/report_invoice.xml',
     ],
