+++ custom_laquage_report/views/custom_laquage.xml
@@ -1,6 +1,6 @@
 <odoo>
-    <template id="report_purchaseorder_document_inherit" inherit_id="studio_customization.studio_report_docume_ec39809c-a15b-4f62-877d-254133881dc8">
-        <xpath expr="//div[contains(@class, 'page')]" position="replace">
+    <template id="report_purchaseorder_document" inherit_id="studio_customization.studio_report_docume_a3fa276b-c49c-488f-80c4-816f3ec0da74" priority="100">
+        <xpath expr="//div[@class='page']" position="replace">
             <div class="page">
                 <!-- En-tête avec logo et informations de l'entreprise -->
                 
@@ -58,8 +58,8 @@
                 <p style="font-weight: bold; font-style: italic; font-size: 15px;">
                     Nous vous remercions de reporter ces références sur vos BL et Factures
                 </p>
-                <p style="margin-bottom: 5px; font-size: 25px;"><strong> RAL (Champs à définir) : </strong><span t-field="doc.partner_ref" /></p>
-                <p style="margin-bottom: 5px; font-size: 25px;"><strong> RICHE EN ZINC (Champs à définir) : </strong><span t-field="doc.partner_ref" /></p>
+                <p style="margin-bottom: 5px; font-size: 25px;"><strong> RAL : </strong><span t-field="doc.x_studio_many2one_field_LCOZX" /></p>
+                <p style="margin-bottom: 5px; font-size: 25px;"><strong> RICHE EN ZINC : </strong><span t-field="doc.x_studio_many2one_field_LCOZX" /></p>
             </div>
              <table class="table table-sm" t-if="doc.state!='done'" name="order.table" style="margin-top: 20px;">
                 <thead>
