+++ fma_custom/__manifest__.py
@@ -8,7 +8,7 @@
     """,
     "summary": "Custom delete button",
     "author": "Odoo PS",
-    "version": "15.0.0.1.0",
+    "version": "17.0.1.0.1",
     "depends": [
         "custom"
     ],
