+++ fma_custom/views/mail_templates.xml
@@ -1,13 +1,12 @@
-<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <template id="mail_notification_paynow_inherit" inherit_id="mail.mail_notification_paynow" name="Invoice: Remove View Invoice Button">
-        <xpath expr="//div[@t-if=&quot;is_online and not record._context.get('proforma')&quot;]" position="replace">
-            <div t-if="is_online and record._context.get('show_view_invoice_button', True) and not record._context.get('proforma')" style="margin: 32px 0px 32px 0px; text-align: center;">
-                <a t-att-href="access_url"
-                    style="background-color: #875A7B; padding: 8px 16px 8px 16px; text-decoration: none; color: #fff; border-radius: 5px; font-size:13px;">
-                    <t t-esc="access_name"/>
+    <template id="mail_notification_layout_inherited" inherit_id="mail.mail_notification_layout">
+        <xpath expr="//td[@t-if='has_button_access']" position="replace">
+            <td t-if="has_button_access and record._context.get('show_view_invoice_button', True)" t-att-style="'border-radius: 3px; text-align: center; background: ' + (company.email_secondary_color or '#875A7B') + ';'">
+                <a t-att-href="button_access['url']" style="font-size: 12px; color: #FFFFFF; display: block; padding: 8px 12px 11px; text-decoration: none !important; font-weight: 400;">
+                    <t t-out="button_access['title']"/>
                 </a>
-            </div>
+            </td>
         </xpath>
     </template>
 </odoo>
