+++ fma_customer_export/models/res_partner.py
@@ -27,110 +27,7 @@ class ResPartner(models.Model):
         """Get customer details for the .txt file based on given fields."""
         content_lines = []
         for partner in partners:
-            line = ['PCC',
-                    'I',
-                    str(partner.part_code_tiers).ljust(9) or '         ',
-                    partner.part_civilite.ljust(5) + partner.name.ljust(30) or '                                   ',
-                    '0',
-                    'N',
-                    'N',
-                    'N',
-                    'N',
-                    '0     ',
-                    'O',
-                    'D',
-                    'N',
-                    'N',
-                    'O',
-                    'N',
-                    ' ',
-                    '         ',
-                    '0 ',
-                    '    ',
-                    'N',
-                    'N',
-                    '0  ',
-                    '   ',
-                    'A41',
-                    '        ',
-                    '       ',
-                    partner.part_civilite.ljust(5) + partner.name.ljust(20) or '                              ',
-                    '                                 ',
-                    '                                ',
-                    '0     ',
-                    '                       ',
-                    'FRA',
-                    str(partner.phone).ljust(20) or '                    ','                    ','                    ',
-                    '    ',
-                    str(partner.x_studio_char_field_G6qIE).ljust(14) or '              ',
-                    '      ',
-                    '              ',
-                    '                         ',
-                    partner.part_commercial.ljust(50) or '',
-                    '             ',
-                    'EUR',
-                    partner.x_studio_mode_de_rglement_1 or '',
-                    '  ',
-                    '                        ',
-                    '                         ',
-                    (str(partner.bank_ids[0].bank_id.name).ljust(5) if (partner.bank_ids and partner.bank_ids[0].bank_id and partner.bank_ids[0].bank_id.name) else '') or '     ',
-                    (str(partner.bank_ids[0].acc_number).ljust(5) if (partner.bank_ids and partner.bank_ids[0].bank_id and partner.bank_ids[0].acc_number) else '') or '     ',
-                    '           ',
-                    '  ',
-                    ' ',
-                    'OOO',
-                    '             ',
-                    'O',
-                    '0000',
-                    '   ',
-                    '   ',
-                    '               ',
-                    'O',
-                    '            ',
-                    'O',
-                    '            ',
-                    '0',
-                    '     ',
-                    'NO1',
-                    '                               ',
-                    str(partner.invoice_ids[0].partner_id.zip).ljust(20) if partner.invoice_ids else str(partner.zip).ljust(20) or '                    ',
-                    '                              ',
-                    '@                             ',
-                    '@                       ',
-                    '@                       ',
-                    '@    ',
-                    '@    ',
-                    '@          ',
-                    '@ ',
-                    '@                             ',
-                    '@                       ',
-                    '@                       ',
-                    '@    ',
-                    '@    ',
-                    '@          ',
-                    '@ ',
-                    '1',
-                    '                                  ',
-                    '                    ',
-                    '                                  ',
-                    '                    ',
-                    '                                  ',
-                    '                    ',
-                    '        ',
-                    '                    ',
-                    'N',
-                    '                                                            ',
-                    '      ',
-                    '@        ',
-                    (str(partner.invoice_ids[0].partner_id.street).ljust(38) if partner.invoice_ids else str(partner.street).ljust(38)) or '                                      ',
-                    (str(partner.invoice_ids[0].partner_id.street2).ljust(38) if (partner.invoice_ids and partner.invoice_ids[0].partner_id.street2) else str(partner.street2).ljust(38)) or '                                       ',
-                    (str(partner.invoice_ids[0].partner_id.state_id.name).ljust(38) if ( partner.invoice_ids and partner.invoice_ids[0].partner_id.state_id.name) else str(partner.state_id.name).ljust(38)) or '                                         ',
-                    str(partner.email).ljust(100) or '                                                                                                                                                   ',
-                    '                            ',
-                    (str(partner.invoice_ids[0].partner_id.city).ljust(26) if partner.invoice_ids else str(partner.city).ljust(26)) or '                             ',
-                    '00001',
-                    'papier ',
-                    '                                                                                    ']
+            line = ['PCC','I',str(partner.part_code_tiers).ljust(9) or '         ',partner.name.ljust(35) or '                                   ','0','N','N','N','0     ','O','D','N','N','O','N',' ','         ','0 ','  ','N','N','0  ','   ','A41','        ','    ',partner.name.ljust(30) or '                              ','                              ','                              ','0     ','                        ','FRA',str(partner.phone).ljust(20) or '                    ','                    ','                    ','    ',str(partner.x_studio_char_field_G6qIE).ljust(14) or '              ','      ','              ','                          ',partner.part_commercial.ljust(50) or '',str(partner.encours_max).ljust(13) or '             ','EUR',partner.x_studio_mode_de_rglement_1 or '','  ','                        ','                         ',((partner.bank_ids[0].bank_id.name).ljust(5) if partner.bank_ids else '') or '     ',(str(partner.bank_ids[0].acc_number).ljust(5) if partner.bank_ids else '') or '     ','           ','  ','','','','   ','          ','','0000','   ','   ','               ','             ','             ','      ','','','','                               ',(str(partner.invoice_ids[0].partner_id.zip).ljust(20) if partner.invoice_ids else partner.zip) or '                    ','                               ','@                             ','@                        ','@                        ','@    ','@    ','@          ','@ ','@                               ','@                        ','@    ','@    ','@          ','@ ','1','                                  ','                    ','                                  ','                    ','                                  ','                    ','NAF     ','                    ','','                                                            ','      ','@        ',(str(partner.invoice_ids[0].partner_id.street).ljust(38) if partner.invoice_ids else str(partner.street).ljust(38)) or '                                      ',(str(partner.invoice_ids[0].partner_id.street2).ljust(38) if partner.invoice_ids else str(partner.street2).ljust(38)) or '                                       ',(str(partner.invoice_ids[0].partner_id.state_id.name).ljust(38) if partner.invoice_ids else str(partner.state_id.name).ljust(38)) or '                                         ',str(partner.email).ljust(100) or '                                                                                                                                                   ','@                            ',(str(partner.invoice_ids[0].partner_id.city).ljust(28) if partner.invoice_ids else str(partner.city).ljust(28)) or '                             ','00001','          ','                                                                                    ']
             content_lines.append(''.join(map(str, line)))
 
         return '\n'.join(content_lines)
@@ -141,7 +38,7 @@ class ResPartner(models.Model):
         for partner in partners:
             # Acknowledge in the chatter
             partner.message_post(
-                body=_("Customer Details files created: <a href='%s' target='_blank'>%s</a>") % (attachment_url, file.name)
+                body=Markup(_("Customer Details files created: <a href='%s' target='_blank'>%s</a>")) % (attachment_url, file.name)
             )
             # Add the file to attachments as well
             partner.attachment_ids = [(4, file.id)]
