+++ fma_customer_outstanding/__manifest__.py
@@ -11,7 +11,7 @@
             Task: 4101724
         """,
     'author': 'Odoo PS',
-    'version': '15.0.0.1.0',
+    'version': '17.0.1.0.1',
     'depends': [
         'contacts',
         'fma_sale_order_custom'
