+++ fma_invoice_export/views/res_config_settings.xml
@@ -5,7 +5,7 @@
         <field name="model">res.config.settings</field>
         <field name="inherit_id" ref="account.res_config_settings_view_form" />
         <field name="arch" type="xml">
-            <xpath expr="//div[@name=('main_currency_setting_container')]" position="after">
+            <xpath expr="//block[@name=('main_currency_setting_container')]" position="after">
                 <h2>Export Invoices</h2>
                 <div class="row mt16 o_settings_container" name="ftp_server_configurations_container">
                     <div class="col-12 col-lg-6 o_setting_box" id="ftp_server_configurations_box">
