+++ fma_invoice_status/__manifest__.py
@@ -10,7 +10,7 @@
             Task: 4115364
         """,
     'author': 'Odoo PS',
-    'version': '15.0.0.1.0',
+    'version': '17.0.1.0.1',
     'depends': [
         'account_accountant'
     ],
