+++ purchase_order_export/views/purchase_order_views.xml
@@ -1,5 +1,4 @@
 <odoo>
-    <!-- Action pour l'exportation de commande -->
     <record id="action_export_order" model="ir.actions.server">
         <field name="name">Export Order</field>
         <field name="model_id" ref="purchase.model_purchase_order"/>
@@ -11,7 +10,6 @@
         </field>
     </record>
 
-    <!-- Héritage de la vue Tree avec ajout des champs -->
     <record id="purchase_order_kpis_tree_inherit" model="ir.ui.view">
         <field name="name">purchase.order.kpis.tree.inherit</field>
         <field name="model">purchase.order</field>
@@ -25,7 +23,6 @@
         </field>
     </record>
 
-    <!-- Héritage de la vue Form avec ajout de champs supplémentaires -->
     <record id="purchase_order_form_inherit" model="ir.ui.view">
         <field name="name">purchase.order.form.inherit</field>
         <field name="model">purchase.order</field>
@@ -35,35 +32,6 @@
             <xpath expr="//div[@name='reminder']" position="after">
                 <field name="shipping_partner_id" domain="[('parent_id', '=', partner_id)]" options="{'no_create': True}"/>
                 <field name="customer_delivery_address"/>
-            </xpath>
-        </field>
-    </record>
-
-    <!-- Ajout de l'onglet Infos Laquage -->
-    <record id="view_purchase_order_form_inherit_with_infos_laquage" model="ir.ui.view">
-        <field name="name">purchase.order.form.inherit.with.infos.laquage</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase.purchase_order_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//notebook" position="inside">
-                <page string="Infos Laquage">
-                    <group>
-                        <field name="laquage_line_ids">
-                            <tree editable="bottom">
-                                <field name="so_repere"/>
-                                <field name="so_qte_commandee"/>
-                                <field name="so_reliquat"/>
-                            </tree>
-                            <form>
-                                <group>
-                                    <field name="so_repere"/>
-                                    <field name="so_qte_commandee"/>
-                                    <field name="so_reliquat"/>
-                                </group>
-                            </form>
-                        </field>
-                    </group>
-                </page>
             </xpath>
         </field>
     </record>
